+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## v1.0.1 (Next Release)
 
- - Added a discovery for Application Perspectives
+ - Added a discovery for application perspectives
+ - Added an action to create a maintenance window
 
 ## v1.0.0